+++ src/django_upgrade/fixers/admin_decorators.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import ast
 from functools import partial
-from typing import Iterable
+from typing import Iterable, Literal
 
 from tokenize_rt import Offset, Token, tokens_to_src
 
@@ -42,15 +42,24 @@ def visit_ClassDef(
 
 # Map from old assigned names to new decorator names, which were changed to be
 # shorter
-ACTION_NAMES = {
-    "short_description": "description",
-    "allowed_permissions": "permissions",
+NAME_MAPS = {
+    "action": {
+        "short_description": "description",
+        "allowed_permissions": "permissions",
+    },
+    "display": {
+        "short_description": "description",
+        "boolean": "boolean",
+        "empty_value_display": "empty_value",
+        "admin_order_field": "ordering",
+    },
 }
 
 
 class FunctionDetails:
-    def __init__(self, node: ast.FunctionDef):
+    def __init__(self, node: ast.FunctionDef, decorator: Literal["action", "display"]):
         self.node = node
+        self.decorator = decorator
         # Discovered attribute assignments to the given function
         self.assignments: dict[str, ast.Assign] = {}
         # Source strings for the values assigned to attributes, used in
@@ -63,8 +72,15 @@ def visit_Module_or_ClassDef(
     node: ast.Module | ast.ClassDef,
     parent: ast.AST,
 ) -> Iterable[tuple[Offset, TokenFunc]]:
-    # Store potential action functions and details of the assigned attributes
-    action_funcs: dict[str, FunctionDetails] = {}
+    # Potential action and display functions to details of assigned attributes
+    funcs: dict[str, FunctionDetails] = {}
+
+    # Display functions take one arg (the model instance), but will also take
+    # self inside classes
+    if isinstance(node, ast.Module):
+        display_func_args = 1
+    else:
+        display_func_args = 2
 
     # Check for 'from django.contrib import admin' from state.from_imports,
     # but also directly when visiting a module. state.from_imports isn’t
@@ -84,33 +100,41 @@ def visit_Module_or_ClassDef(
             )
         ):
             admin_imported = True
-        elif (
-            isinstance(subnode, ast.FunctionDef)
-            # Django calls action functions with exactly three arguments,
-            # positionally (modeladmin, request, queryset)
-            and (len(subnode.args.posonlyargs) + len(subnode.args.args)) == 3
-            and len(subnode.args.kwonlyargs) == 0
-            # TODO: check that no admin.action decorator already applied
-        ):
-            action_funcs[subnode.name] = FunctionDetails(subnode)
+        elif isinstance(subnode, ast.FunctionDef):
+            if (
+                # Django calls action functions with exactly three arguments,
+                # positionally (modeladmin, request, queryset)
+                (len(subnode.args.posonlyargs) + len(subnode.args.args)) == 3
+                and len(subnode.args.kwonlyargs) == 0
+                # TODO: check that no admin.action decorator already applied
+            ):
+                funcs[subnode.name] = FunctionDetails(subnode, "action")
+            elif (
+                (len(subnode.args.posonlyargs) + len(subnode.args.args))
+                == display_func_args
+                and len(subnode.args.kwonlyargs) == 0
+                # TODO: check that no admin.display decorator already applied
+            ):
+                funcs[subnode.name] = FunctionDetails(subnode, "display")
         elif (
             isinstance(subnode, ast.Assign)
             and len(subnode.targets) == 1
             and isinstance((target := subnode.targets[0]), ast.Attribute)
             and isinstance(target.value, ast.Name)
-            and target.value.id in action_funcs
-            and target.attr in ACTION_NAMES
+            and target.value.id in funcs
+            and target.attr in NAME_MAPS[funcs[target.value.id].decorator]
         ):
-            new_name = ACTION_NAMES[target.attr]
-            action_funcs[target.value.id].assignments[new_name] = subnode
+            names = NAME_MAPS[funcs[target.value.id].decorator]
+            new_name = names[target.attr]
+            funcs[target.value.id].assignments[new_name] = subnode
 
     if not admin_imported:
         return
 
-    for name, funcdetails in action_funcs.items():
+    for name, funcdetails in funcs.items():
         if funcdetails.assignments:
             yield ast_start_offset(funcdetails.node), partial(
-                decorate_action_function, funcdetails=funcdetails
+                decorate_function, funcdetails=funcdetails
             )
             for name, assignnode in funcdetails.assignments.items():
                 yield ast_start_offset(assignnode), partial(erase_node, node=assignnode)
@@ -122,12 +146,13 @@ def visit_Module_or_ClassDef(
                 )
 
 
-def decorate_action_function(
+def decorate_function(
     tokens: list[Token], i: int, *, funcdetails: FunctionDetails
 ) -> None:
     j, indent = extract_indent(tokens, i)
-    dec_src = f"{indent}@admin.action(\n"
-    for name in ACTION_NAMES.values():  # Use predefined order
+    dec_src = f"{indent}@admin.{funcdetails.decorator}(\n"
+    names = NAME_MAPS[funcdetails.decorator]
+    for name in names.values():  # Use predefined order
         if name not in funcdetails.values:
             continue
         source = funcdetails.values[name]